+++ modules/kubernetes/aks-core/modules.tf
@@ -452,6 +452,7 @@ module "prometheus" {
   aad_pod_identity_enabled                 = var.aad_pod_identity_enabled
   azad_kube_proxy_enabled                  = var.azad_kube_proxy_enabled
   starboard_enabled                        = var.starboard_enabled
+  vpa_enabled                              = var.vpa_enabled
 }
 
 # starboard
@@ -467,6 +468,19 @@ module "starboard" {
   source = "../../kubernetes/starboard"
 
   cloud_provider = "azure"
+}
+
+# vpa
+module "vpa" {
+  depends_on = [module.opa_gatekeeper, module.prometheus]
+
+  for_each = {
+    for s in ["vpa"] :
+    s => s
+    if var.vpa_enabled
+  }
+
+  source = "../../kubernetes/vpa"
 }
 
 module "new_relic" {