+++ modules/kubernetes/prometheus/main.tf
@@ -42,7 +42,9 @@ resource "helm_release" "prometheus" {
   namespace   = kubernetes_namespace.this.metadata[0].name
   version     = "30.0.0"
   max_history = 3
-  values      = [templatefile("${path.module}/templates/values.yaml.tpl", {})]
+  values = [templatefile("${path.module}/templates/values.yaml.tpl", {
+    vpa_enabled = var.vpa_enabled,
+  })]
 }
 
 # EKS will not install metrics server out of the box so it has to be added.